moves hana-saml to lib folder
