+++ lib/hana-saml.js
