+++ package.json
@@ -17,5 +17,9 @@
   "bugs": {
     "url": "https://github.com/gregorwolf/hanatrial-auth-proxy/issues"
   },
-  "homepage": "https://github.com/gregorwolf/hanatrial-auth-proxy"
+  "homepage": "https://github.com/gregorwolf/hanatrial-auth-proxy",
+  "dependencies": {
+    "cheerio": "^0.19.0",
+    "request": "^2.54.0"
+  }
 }
