+++ saml-client.js
@@ -1,5 +1,5 @@
 var https = require('https');
-var hanaSaml = require('./hana-saml');
+var hanaSaml = require('./lib/hana-saml');
 
 // Grab the config file
 var config;
